Move processStatesMutliSelectOptions in the new ui services directory
