+++ ui/main/src/app/modules/feed/components/card-list/filters/feed-filter/feed-filter.component.ts
@@ -21,7 +21,7 @@ import {OpfabStore} from 'app/business/store/opfabStore';
 import {MultiSelect, MultiSelectOption} from 'app/modules/share/multi-select/model/MultiSelect';
 import {ProcessesService} from '@ofServices/processes/ProcessesService';
 import {UsersService} from '@ofServices/users/UsersService';
-import {ProcessStatesMultiSelectOptionsService} from 'app/business/services/process-states-multi-select-options.service';
+import {ProcessStatesMultiSelectOptionsService} from '@ofServices/processStatesMultiSelectOptions/ProcessStatesMultiSelectOptionsService';
 import {TranslateModule} from '@ngx-translate/core';
 import {NgIf} from '@angular/common';
 import {FeedSortComponent} from '../feed-sort/feed-sort.component';
